# MRG: In Report.add_raw(), re-use the existing figure when creating butterfly plots of slices