Performance changes **EDIT I've updated the numbers after adding dde9732**

```
main: 12.3 s ± 294 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
pr:   6.98 s ± 429 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Because we're re-using the same figure, I could also get rid of the explicit `scalings` pre-calculation.


MWE:
```python
# %%
from pathlib import Path
import mne


sample_dir = Path(mne.datasets.sample.data_path())
sample_fname = sample_dir / 'MEG' / 'sample' / 'sample_audvis_raw.fif'

raw = mne.io.read_raw_fif(sample_fname, preload=True)
raw.crop(tmax=60)
r = mne.Report()

# %%
%timeit r.add_raw(raw=raw, title='foo')
```

x-ref #10107