@hoechenberger 2021-12-09
`_render_slider()` now has a new `imgs` kwarg in addition to the old `figs`, which allows to pass images directly instead of only MPL figures. This allows us to take "snapshots" of the same figure in the `for` loop without having to create a copy first.